Add copy/paste for Connector

+++ gaphor/UML/components/connectorconnect.py
@@ -107,9 +107,10 @@ class ConnectorConnectBase(BaseConnector):
          component
             Component item.
         """
-        p = component.subject.ownedPort[0]
-        p.unlink()
-        connector.subject = None
+        if component and component.subject:
+            p = component.subject.ownedPort[0]
+            p.unlink()
+            connector.subject = None
 
     def allow(self, handle, port):
         iface = self.element
